Help command replaces -help. help, map commands gets generated.
