+++ private/scripts/cmdsMap.js
@@ -95,7 +95,8 @@ commands.map = {
     off: { description: 'Turn off map' },
     zoomin: { description: 'Increase zoom' },
     zoomout: { description: 'Decrease zoom' },
-    locate: { description: 'Locate person or place' },
+    user: { description: 'Locate user', next: {} },
+    location: { description: 'Locate location', next: {} },
     info: { description: 'Show information for the location' },
   },
   commandName: 'map',
